requirements.txt
Flask==1.1.1
flask_cors==3.0.8
loader==2017.9.11
python-dotenv==0.12.0
Werkzeug==1.0.0
